Load scripts are separated. Docker compose without .env
